examples/llm_prompt_optimization/wilds_experiment/requirements.txt
wilds>=2.0.0
numpy>=1.21.0
matplotlib>=3.5.0
jsonlines>=3.0.0
openai>=1.0.0
pyyaml>=6.0
tqdm>=4.62.0
scikit-learn>=1.0.0
torch>=1.9.0
pandas>=1.3.0

